fixed foreign keys

+++ database/migrations/2022_04_28_000027_create_job_applicants_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
     {
         Schema::create('job_applicants', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('job_id');
+            $table->unsignedBigInteger('job_id');
             $table->string('firstname');
             $table->string('lastname');
             $table->string('email');
